move test_verify_backingstore_uses_rgw to test_provider_client.py
